New utils file to make the Column.tsx clean
